src/Web/Coupon_Django_Site/requirements.txt
Django==5.0.3
psycopg2_binary==2.9.9
